agent/requirements.txt
browser-use==0.1.40
langchain==0.3.14
groq
langchain-groq
python-dotenv
tenacity
playwright
